# swagger-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "REST API for comments app",
    "description": "dZENcode test assignment API documentation",
    "contact": {
      "name": "Oleksandr Kochenko",
      "url": "https://oleksandrkochenko.github.io/CV/",
      "email": "kochenko.fop@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://dzencode-test-ryk0.onrender.com"
    },
    {
      "url": "http://localhost:3000/"
    }
  ],
...
  "paths": {
...
    "/auth/current": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "Check user ligged in",
        "operationId": "current",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_2"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Unauthorized": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "error message",
            "example": "Unauthorized request"
          }
        }
      },
      "ServerError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Server error"
          }
        }
      },
...
      "inline_response_200_2": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "operation message",
            "example": "User is current"
          },
          "user": {
            "$ref": "#/components/schemas/inline_response_200_2_user"
          }
        }
      },
...
      "inline_response_200_2_user": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "current user id",
            "example": "65538751b374715901fe334a"
          },
          "name": {
            "type": "string",
            "description": "current user name",
            "example": "Peter Peterson"
          },
          "email": {
            "type": "string",
            "description": "current user email",
            "example": "batman2@mail.com"
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}